% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_10.R
\name{get_tab_10}
\alias{get_tab_10}
\title{get_tab_10}
\usage{
get_tab_10(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo II de los datos descargados desde la pagina del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Ordena los datos del Cuadro N 10 del Tomo II de los Resultados del Censo Nacional de 2017.

Esta funcion permite organizar los datos del Cuadro N 10 del Tomo II de los
Resultados del Censo Nacional de 2017, el cual tiene el siguiente titulo:
"POBLACION CENSADA, POR ALGUNA DIFICULTAD O LIMITACION PERMANENTE, SEGUN PROVINCIA,
DISTRITO, AREA URBANA Y RURAL, SEXO Y GRUPOS DE EDAD".
}
