% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_1e.R
\name{get_tab_1_edu}
\alias{get_tab_1_edu}
\title{get_tab_1_edu}
\usage{
get_tab_1_edu(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo III de los datos descargados desde la pagina del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Esta funcion permite organizar los datos del Cuadro 1  relacionado con el tema de educacion
del Tomo III de los Resultados del Censo Nacional de 2017, el cual tiene el siguiente titulo:
"POBLACION CENSADA DE 3  Y MAS ANOS DE EDAD, POR GRUPOS DE EDAD, SEGUN PROVINCIA, DISTRITO, AREA URBANA Y RURAL,
SEXO E IDIOMA O LENGUA MATERNA CON EL QUE APRENDIO A HABLAR EN SU NINEZ"
}
