% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_3e.R
\name{get_tab_3_edu}
\alias{get_tab_3_edu}
\title{get_tab_3_edu}
\usage{
get_tab_3_edu(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo VI de los datos descargados desde la pagina del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Esta funcion permite organizar los datos del Cuadro 3 relacionado con el tema de educacion
del Tomo VI de los Resultados del Censo Nacional de 2017, el cual tiene el siguiente titulo:
"POBLACION CENSADA DE 3 Y MAS ANOS DE EDAD, POR GRUPOS DE EDAD, SEGUN PROVINCIA, DISTRITO,
AREA URBANA Y RURAL, SEXO Y NIVEL EDUCATIVO ALCANZADO"
}
