% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_3s.R
\name{get_tab_3_salud}
\alias{get_tab_3_salud}
\title{get_tab_3_salud}
\usage{
get_tab_3_salud(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo II de los datos descargados desde la página del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Esta función permite organizar los datos del Cuadro Nº 3  relacionado con el tema de salud
del Tomo II de los Resultados del Censo Nacional de 2017, el cual tiene el siguiente título:
"POBLACIÓN CENSADA EN VIVIENDAS PARTICULARES, POR AFILIACIÓN A ALGÚN TIPO DE SEGURO DE SALUD,
SEGÚN PROVINCIA, DISTRITO, ÁREA URBANA Y RURAL, SEXO Y RELACIÓN DE PARENTESCO
CON EL JEFE O JEFA DEL HOGAR".
}
