% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_4.R
\name{get_tab_4}
\alias{get_tab_4}
\title{Procesar datos del cuadro de población censada en viviendas particulares}
\usage{
get_tab_4(file, sheet, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel que contiene los datos.}

\item{sheet}{Número de la hoja en el archivo Excel que contiene los datos.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos procesados.
}
\description{
Esta función lee y procesa los datos del cuadro de población censada en viviendas particulares,
según provincia, distrito, área urbana y rural, sexo y relación de parentesco con el jefe o jefa del hogar.
}
\examples{
\dontrun{
df <-  get_tab_4("rawdata/08TOMO_01.xlsx", sheet = 4)
}
}
