% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_5.R
\name{get_tab_5}
\alias{get_tab_5}
\title{Procesar datos del cuadro de población censada por tenencia de documento de identidad}
\usage{
get_tab_5(file, sheet, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel que contiene los datos.}

\item{sheet}{Número de la hoja en el archivo Excel que contiene los datos.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos procesados.
}
\description{
Esta función lee y procesa los datos del cuadro de población censada por tenencia de algún tipo de documento de identidad,
según provincia, distrito, área urbana y rural, grupos de edad y sexo.
}
\examples{
\dontrun{
df <- process_identity_data("rawdata/08TOMO_01.xlsx", sheet = 5, )
}
}
