% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_5s.R
\name{get_tab_5_salud}
\alias{get_tab_5_salud}
\title{get_tab_5_salud}
\usage{
get_tab_5_salud(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo III de los datos descargados desde la pagina del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Esta funcion permite organizar los datos del Cuadro 5  relacionado con el tema de salud
del Tomo III de los Resultados del Censo Nacional de 2017, el cual tiene el siguiente título:
" POBLACION CENSADA DE 3 Y MAS ANOS DE EDAD, POR AFILIACION A ALGUN TIPO DE SEGURO DE SALUD,
SEGUN PROVINCIA, DISTRITO, AREA URBANA Y RURAL, SEXO Y NIVEL EDUCATIVO ALCANZADO".
}
