% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_7.R
\name{get_tab_7}
\alias{get_tab_7}
\title{Procesar datos del cuadro de poblacion censada por sexo y grupos de edad}
\usage{
get_tab_7(file, sheet, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel que contiene los datos.}

\item{sheet}{Numero de la hoja en el archivo Excel que contiene los datos.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos procesados.
}
\description{
Esta funcion lee y procesa los datos del cuadro de poblacion censada por sexo y grupos de edad,
según lugar de residencia permanente dentro y fuera del país.
}
\examples{
\dontrun{
df <- get_tab_7("rawdata/08TOMO_02.xlsx", sheet = 2, skip = 4)
}
}
