% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_9.R
\name{get_tab_9}
\alias{get_tab_9}
\title{Procesar datos del cuadro de población censada por sexo y grupos de edad según lugar de residencia de la madre cuando usted nació}
\usage{
get_tab_9(file, sheet, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel que contiene los datos.}

\item{sheet}{Número de la hoja en el archivo Excel que contiene los datos.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos procesados.
}
\description{
Esta función lee y procesa los datos del cuadro de población censada por sexo y grupos de edad, según lugar de residencia de la madre cuando usted nació,
dentro y fuera del país.
}
