% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_1.R
\name{get_tab_1}
\alias{get_tab_1}
\title{read_pob_tab_1}
\usage{
get_tab_1(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble que muestra datos demográficos desglosados por:
departamento, provincia, distrito, distribución (urbano/rural), sexo,
edad y población.
}
\description{
Los datos son tomados desde la plataforma del INSTITUTO NACIONAL DE ESTADÍSTICA E INFORMÁTICA - INEI
https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/
}
\details{
Lee la información del cuadro "POBLACIÓN CENSADA, POR ÁREA URBANA Y RURAL;
Y SEXO, SEGÚN PROVINCIA, DISTRITO Y EDADES SIMPLES" y genera una tabla en
formato tidy.

Esta función toma como argumentos el archivo donde se encuentra el cuadro y
el nombre del departamento.
}
