% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_10.R
\name{get_tab_10}
\alias{get_tab_10}
\title{Obtener datos del cuadro de población censada por alguna dificultad o limitación permanente}
\usage{
get_tab_10(file, sheet = 5, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel que contiene los datos.}

\item{sheet}{Número de la hoja en el archivo Excel que contiene los datos.}

\item{dep_name}{Nombre del departamento.}
}
\value{
Un tibble con los datos procesados.
}
\description{
Esta función lee y procesa los datos del cuadro de población censada por alguna dificultad o limitación permanente,
según provincia, distrito, área urbana y rural, sexo y grupos de edad.
}
