% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_3.R
\name{get_tab_3}
\alias{get_tab_3}
\title{Ordenar datos del cuadro de poblacion censada en viviendas particulares}
\usage{
get_tab_3(file, sheet, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo de Excel que contiene los datos.}

\item{sheet}{Numero de la hoja de Excel que contiene los datos del cuadro.}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble ordenado con la informacion procesada.
}
\description{
Esta función procesa y ordena adecuadamente los datos obtenidos del cuadro
"POBLACION CENSADA EN VIVIENDAS PARTICULARES, POR GRUPOS DE EDAD",
presentado en el Censo Poblacional del Peru 2017.
}
\examples{
\dontrun{
df <- read_pob_vivienda("rawdata/08TOMO_01.xlsx", sheet = 3)
}
}
