% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_7.R
\name{get_tab_7}
\alias{get_tab_7}
\title{get_tab_7}
\usage{
get_tab_7(file, dep_name = NULL)
}
\arguments{
\item{file}{Ruta del archivo Excel del Tomo II de los datos descargados desde la página del INEI
(https://censo2017.inei.gob.pe/resultados-definitivos-de-los-censos-nacionales-2017/).}

\item{dep_name}{Nombre del departamento al que pertenecen los datos.}
}
\value{
Un tibble con los datos ordenados en formato largo.
}
\description{
Ordena los datos del Cuadro Nº 7 del Tomo II de los Resultados del Censo Nacional de 2017.

Esta función permite organizar los datos del Cuadro Nº 7 del Tomo II de los
Resultados del Censo Nacional de 2017, el cual tiene el siguiente título:
"POBLACIÓN CENSADA, POR SEXO Y GRUPOS DE EDAD, SEGÚN LUGAR DE RESIDENCIA PERMANENTE
DENTRO Y FUERA DEL PAÍS".
}
\examples{
\dontrun{
df <- get_tab_7("rawdata/08TOMO_02.xlsx", dep_name = "CUSCO")
}
}
